requirements.txt
Django>=4.0
djangorestframework>=3.12
drf-yasg>=1.20.0
djangorestframework-simplejwt>=5.0
psycopg2-binary>=2.9
python-dotenv>=0.20.0
django-cors-headers>=3.10